requirements.txt
amqp==5.3.1
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.3
billiard==4.2.1
binaryornot==0.4.4
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.6.0
cron-descriptor==1.4.5
cryptography==44.0.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.0.13
djangorestframework==3.15.2
executing==2.1.0
fido2==1.2.0
filelock==3.18.0
httpie==3.2.4
humanize==4.12.2
identify==2.6.9
idna==3.10
inflection==0.5.1
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
nodeenv==1.9.1
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.6
pre_commit==4.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psycopg==3.2.9
psycopg-binary==3.2.9
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
text-unidecode==1.3
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
virtualenv==20.29.3
wcwidth==0.2.13
